Label-less ADTs make the representation so much cleaner
For this model language we don't really need to keep track of any
labels, and we can assume that each label _is_ just a number tracking
the position of the con/branch.

This implies that case branches must be in order of constructors
declaration, but that is fine for the model language.