Some progress on proving substitution on terms
I guess that's about as far as I can go without stumbling upon the
problems of expressing variable bindings the way I do. The rest of the
clauses are either trivial (hence boring) or require me to rework the
way I'm treating variables.

This stringly-typed thing had a smell from the start, I knew I'll get
hit with it.