Remove a clearly incorrect theorem
Suppose that
τ∈Γ-⇒-Γ⊢τ-leq : (Γok : Γ ok)
              → (∈-prf : τ ∈ Γ at ι)
              → size-twf (τ∈Γ-⇒-Γ⊢τ Γok ∈-prf) ≤ size-ok Γok
holds.
That is, for any `Γ ok` and a proof of `τ ∈ Γ` we get a derivation of `Γ ⊢ τʻ
and that derivation's size is less than or equal to the size of `Γ ok`.
But, we've proven (Γ⊢τ-⇒-Γok-smaller) that if we go from `Γ ⊢ τ` to `Γ ok`,
then that derivation is strictly smaller than the derivation of `Γ ⊢ τ`,
thus strictly smaller than itself.