Start proving to-sub' that also expands to an arbitrary supertype
And here I need <: to be transitive, for which I need narrowing on
algorithmic derivation to hold, which doesn't really hold, which sucks.

This is really interesting, since I started this "iso" exercise just to
avoid proving narrowing for the preservation theorem, and I thought
about proving preservation via conversion to the declarative typing and
then back.

It's almost as if this thing just doesn't want to be proven. Or, more
likely, it's just me being bad at proving things.