Curiously, reverting to the Minimal type system for narrowing makes things simpler
With the Enriched type system, we need to also do narrowing on the
witnesses of the derivations having well-formed types. This leads to
nasty non-structural recursion, so it seems to be easier to prove
narrowing for the minimal type system first.

Of course, we'll also need to prove that the minimal type system implies
the enriched one, but we'd need to prove this anyway.