Oh crap, looks like I won't be able to prove substitution directly for the enriched system
Both systems require transitivity of subtyping relations to collapse
any possible nested T-Sub rules after the substitution. In turn,
transitivity of subtyping relations for ST-Arr requires narrowing of the
"higher" relation, which, in case of the enriched system, also pulls in
narrowing for the type well-formedness witnesses that it carries.