The Great Cleanse: remove the arrow type δ from T-Abs premises
The well-formedness of the arrow type is somewhat easily derivable from
the well-typedness of the body of the λ-abstraction (as the agrement
lemmas now show).

Removing the extra premise simplifies the lemmas that need to _produce_
new typing derivations, like the narrowing and substitution lemmas.

Why do we need all this?

Right now, one of the T-App's premises says the result type is
well-formed. This is good in some circumstances since it allows proving
the term ⇒ type agreement lemma without having to prove
the substitution lemma. The problem is that we will still need the
substitution lemma if we want to prove preservation.

Now, to prove substitution lemma, we in particular need to prove that
the result type of a T-App after substitution is well-typed.
We could have used substitution lemma for this recursively, but the
problem is that the recursion is no longer structural, so the
termination checker is no longer happy. I also wasn't able to come up
with a good size measure reasonably quickly, so it seems like it's
easier to just remove this premise and have it propositionally instead.