The Great Cleanse: remove the type δ witness from T-Sub
Otherwise, in sub-Γ⊢ε⦂τ / T-App / t-sub case, we'd need the witness of
the domain of ε₁δ' being well-formed, to pass it to as-sub'.
Yet, the only way to obtain this witness would be to apply the
term ⇒ type agreement lemma, but we're trying to make sure that
the substitution lemma is provable without depending on it it.

See 2cdd0985b5a33d1f1a525f1358936be8a0cb841a for more motivation why
we're trying to make the substitution lemma provable without the
agreement lemma.