Change SCase syntax to mention the ADT and result types
This way, we can have rules ending with T-Sub in the scrutinee and
branches of the T-Case rules, while still having uniqueness of typing.

Substitution lemma (and evaluation in general) can refine the type of
the scrutinee and the branches. So, we could either refine the type of
the whole T-Case construct, or preserve the types via T-Sub.

If we were to refine, then we'd need parallel narrowing lemma for
branches, which is annoying but doable, but we'd also need to find joins
of the subtypes, which looks to be undecidable for a type system this
powerful (but we could probably have swept the undecidability under the
rug of the oracle… but that's hairy).

So, we'll need to end the scrutinee and the branches with T-Sub. But
this means that, to preserve the uniqueness of typing, we'd need to
preserve the expected types of each in the syntax.